Remove Redundant Transcripts

Remove Redundant Transcripts

Usage

remove_redundants(
  data_matrix,
  enst_ensp_sequences,
  mane_select = NULL,
  peptide_lengths = NULL
)

Arguments

data_matrix

Bulk RNA Seq expression matrix from samples, columns are samples and first row is gene ids, second row is transcript ids and the rest of the rows are samples

enst_ensp_sequences

amino acid sequence of each ensp

mane_select

mane transcript list is needed to decide which transcript will be kept

Value

dataframe